Add new test for abbreviations
